revert preprocess.py
